Removed legacy code.
