--- src/direct/types.ts
@@ -1,34 +1,0 @@
-import { BigNumber } from '../utils/bignumber';
-
-export interface Pool {
-    id: string;
-    balanceIn: BigNumber;
-    balanceOut: BigNumber;
-    weightIn: BigNumber;
-    weightOut: BigNumber;
-    swapFee: BigNumber;
-    spotPrice?: BigNumber;
-    slippage?: BigNumber;
-    limitAmount?: BigNumber;
-}
-
-export interface SwapAmount {
-    pool: string;
-    amount: BigNumber;
-}
-
-export interface EffectivePrice {
-    price?: BigNumber;
-    id?: string;
-    maxAmount?: string;
-    swap?: string[];
-    amounts?: BigNumber[];
-    bestPools?: string[];
-}
-
-export type Swap = {
-    pool: string;
-    tokenInParam: string;
-    tokenOutParam: string;
-    maxPrice: string;
-};
