+++ src/helpers.ts
@@ -24,10 +24,6 @@ import {
 } from './bmath';
 const disabledTokensDefault = require('./disabled-tokens.json');
 
-export function toChecksum(address) {
-    return utils.getAddress(address);
-}
-
 export function getLimitAmountSwap(
     poolPairData: PoolPairData,
     swapType: string
@@ -500,271 +496,6 @@ function filterPoolsWithoutToken(pools, token) {
     return OutputPools;
 }
 
-// Inputs:
-// - pools: All pools that contain a token
-// - token: Token for which we are looking for pairs
-// Outputs:
-// - tokens: Set (without duplicate elements) of all tokens that pair with token
-function getTokensPairedToTokenWithinPools(pools, token) {
-    let found;
-    let tokens = new Set();
-    for (let i in pools) {
-        found = false;
-        for (let k = 0; k < pools[i].tokensList.length; k++) {
-            if (
-                utils.getAddress(pools[i].tokensList[k]) !=
-                    utils.getAddress(token) &&
-                pools[i].tokens.find(
-                    t =>
-                        utils.getAddress(t.address) ===
-                        utils.getAddress(pools[i].tokensList[k])
-                ).balance != 0
-            ) {
-                tokens.add(pools[i].tokensList[k]);
-            }
-        }
-    }
-    return tokens;
-}
-
-// Returns two arrays
-// First array contains all tokens in direct pools containing tokenIn
-// Second array contains all tokens in multi-hop pools containing tokenIn
-export function getTokenPairsMultiHop(token: string, poolsTokensListSet: any) {
-    let poolsWithToken = [];
-    let poolsWithoutToken = [];
-
-    let directTokenPairsSet = new Set();
-
-    // If pool contains token add all its tokens to direct list
-    poolsTokensListSet.forEach((poolTokenList, index) => {
-        if (poolTokenList.includes(token)) {
-            poolsWithToken.push(poolTokenList);
-        } else {
-            poolsWithoutToken.push(poolTokenList);
-        }
-    });
-
-    directTokenPairsSet = new Set([].concat(...poolsWithToken));
-
-    let multihopTokenPools = [];
-    let multihopTokenPairsSet = new Set();
-
-    poolsWithoutToken.forEach((pool, index) => {
-        let intersection = [...pool].filter(x =>
-            [...directTokenPairsSet].includes(x)
-        );
-        if (intersection.length != 0) {
-            multihopTokenPools.push(pool);
-        }
-    });
-
-    multihopTokenPairsSet = new Set([].concat(...multihopTokenPools));
-    let allTokenPairsSet = new Set();
-    allTokenPairsSet = new Set([
-        ...directTokenPairsSet,
-        ...multihopTokenPairsSet,
-    ]);
-
-    let directTokenPairs = [...directTokenPairsSet];
-    let allTokenPairs = [...allTokenPairsSet];
-    return [directTokenPairs, allTokenPairs];
-}
-
-// Filters all pools data to find pools that have both tokens
-// TODO: Check for balance > 0
-export function filterPoolsWithTokensDirect(
-    allPools: Pool[], // The complete information of the pools
-    tokenIn: string,
-    tokenOut: string,
-    disabledOptions: DisabledOptions = { isOverRide: false, disabledTokens: [] }
-): PoolDictionary {
-    let poolsWithTokens: PoolDictionary = {};
-    // If pool contains token add all its tokens to direct list
-
-    let disabledTokens = disabledTokensDefault.tokens;
-    if (disabledOptions.isOverRide)
-        disabledTokens = disabledOptions.disabledTokens;
-
-    allPools.forEach(pool => {
-        let tokenListSet = new Set(pool.tokensList);
-        disabledTokens.forEach(token => tokenListSet.delete(token.address));
-
-        if (tokenListSet.has(tokenIn) && tokenListSet.has(tokenOut)) {
-            poolsWithTokens[pool.id] = pool;
-        }
-    });
-
-    return poolsWithTokens;
-}
-
-// Returns two pool lists. One with all pools containing tokenOne and not tokenTwo and one with tokenTwo not tokenOn.
-export function filterPoolsWithoutMutualTokens(
-    allPools: Pool[],
-    tokenOne: string,
-    tokenTwo: string,
-    disabledTokens: DisabledToken[] = []
-): [PoolDictionary, Set<string>, PoolDictionary, Set<string>] {
-    let tokenOnePools: PoolDictionary = {};
-    let tokenTwoPools: PoolDictionary = {};
-    let tokenOnePairedTokens: Set<string> = new Set();
-    let tokenTwoPairedTokens: Set<string> = new Set();
-
-    allPools.forEach(pool => {
-        let poolTokensSET = new Set(pool.tokensList);
-        disabledTokens.forEach(token => poolTokensSET.delete(token.address));
-        if (poolTokensSET.size < 2) {
-            return;
-        }
-
-        let containsTokenOne = poolTokensSET.has(tokenOne);
-        let containsTokenTwo = poolTokensSET.has(tokenTwo);
-
-        if (containsTokenOne && !containsTokenTwo) {
-            tokenOnePairedTokens = new Set([
-                ...tokenOnePairedTokens,
-                ...poolTokensSET,
-            ]);
-            tokenOnePools[pool.id] = pool;
-        } else if (!containsTokenOne && containsTokenTwo) {
-            tokenTwoPairedTokens = new Set([
-                ...tokenTwoPairedTokens,
-                ...poolTokensSET,
-            ]);
-            tokenTwoPools[pool.id] = pool;
-        }
-    });
-
-    return [
-        tokenOnePools,
-        tokenOnePairedTokens,
-        tokenTwoPools,
-        tokenTwoPairedTokens,
-    ];
-}
-
-// Replacing getMultihopPoolsWithTokens
-export function filterPoolsWithTokensMultihop(
-    allPools: Pool[], // Just the list of pool tokens
-    tokenIn: string,
-    tokenOut: string,
-    disabledOptions: DisabledOptions = { isOverRide: false, disabledTokens: [] }
-): [Pool[], Pool[], string[]] {
-    //// Multi-hop trades: we find the best pools that connect tokenIn and tokenOut through a multi-hop (intermediate) token
-    // First: we get all tokens that can be used to be traded with tokenIn excluding
-    // tokens that are in pools that already contain tokenOut (in which case multi-hop is not necessary)
-    let poolsTokenInNoTokenOut: PoolDictionary;
-    let tokenInHopTokens: Set<string>;
-    let poolsTokenOutNoTokenIn: PoolDictionary;
-    let tokenOutHopTokens: Set<string>;
-
-    let disabledTokens = disabledTokensDefault.tokens;
-    if (disabledOptions.isOverRide)
-        disabledTokens = disabledOptions.disabledTokens;
-
-    // STOPPED HERE: poolsTokenInNoTokenOut NEEDS
-    [
-        poolsTokenInNoTokenOut,
-        tokenInHopTokens,
-        poolsTokenOutNoTokenIn,
-        tokenOutHopTokens,
-    ] = filterPoolsWithoutMutualTokens(
-        allPools,
-        tokenIn,
-        tokenOut,
-        disabledTokens
-    );
-
-    // Third: we find the intersection of the two previous sets so we can trade tokenIn for tokenOut with 1 multi-hop
-    const hopTokensSet = [...tokenInHopTokens].filter(x =>
-        tokenOutHopTokens.has(x)
-    );
-
-    // Transform set into Array
-    const hopTokens: string[] = [...hopTokensSet];
-    // console.log(hopTokens);
-
-    // Find the most liquid pool for each pair (tokenIn -> hopToken). We store an object in the form:
-    // mostLiquidPoolsFirstHop = {hopToken1: mostLiquidPool, hopToken2: mostLiquidPool, ... , hopTokenN: mostLiquidPool}
-    // Here we could query subgraph for all pools with pair (tokenIn -> hopToken), but to
-    // minimize subgraph calls we loop through poolsTokenInNoTokenOut, and check the liquidity
-    // only for those that have hopToken
-    let mostLiquidPoolsFirstHop: Pool[] = [];
-    for (let i = 0; i < hopTokens.length; i++) {
-        let highestNormalizedLiquidity = bnum(0); // Aux variable to find pool with most liquidity for pair (tokenIn -> hopToken)
-        let highestNormalizedLiquidityPoolId; // Aux variable to find pool with most liquidity for pair (tokenIn -> hopToken)
-        for (let k in poolsTokenInNoTokenOut) {
-            // If this pool has hopTokens[i] calculate its normalized liquidity
-            if (
-                new Set(poolsTokenInNoTokenOut[k].tokensList).has(hopTokens[i])
-            ) {
-                let normalizedLiquidity = getNormalizedLiquidity(
-                    parsePoolPairData(
-                        poolsTokenInNoTokenOut[k],
-                        tokenIn,
-                        hopTokens[i].toString()
-                    )
-                );
-
-                if (
-                    normalizedLiquidity.isGreaterThanOrEqualTo(
-                        // Cannot be strictly greater otherwise
-                        // highestNormalizedLiquidityPoolId = 0 if hopTokens[i] balance is 0 in this pool.
-                        highestNormalizedLiquidity
-                    )
-                ) {
-                    highestNormalizedLiquidity = normalizedLiquidity;
-                    highestNormalizedLiquidityPoolId = k;
-                }
-            }
-        }
-        mostLiquidPoolsFirstHop[i] =
-            poolsTokenInNoTokenOut[highestNormalizedLiquidityPoolId];
-        // console.log(highestNormalizedLiquidity)
-        // console.log(mostLiquidPoolsFirstHop)
-    }
-
-    // console.log('mostLiquidPoolsFirstHop');
-    // console.log(mostLiquidPoolsFirstHop);
-
-    // Now similarly find the most liquid pool for each pair (hopToken -> tokenOut)
-    let mostLiquidPoolsSecondHop: Pool[] = [];
-    for (let i = 0; i < hopTokens.length; i++) {
-        let highestNormalizedLiquidity = bnum(0); // Aux variable to find pool with most liquidity for pair (tokenIn -> hopToken)
-        let highestNormalizedLiquidityPoolId; // Aux variable to find pool with most liquidity for pair (tokenIn -> hopToken)
-        for (let k in poolsTokenOutNoTokenIn) {
-            // If this pool has hopTokens[i] calculate its normalized liquidity
-            if (
-                new Set(poolsTokenOutNoTokenIn[k].tokensList).has(hopTokens[i])
-            ) {
-                let normalizedLiquidity = getNormalizedLiquidity(
-                    parsePoolPairData(
-                        poolsTokenOutNoTokenIn[k],
-                        hopTokens[i].toString(),
-                        tokenOut
-                    )
-                );
-
-                if (
-                    normalizedLiquidity.isGreaterThanOrEqualTo(
-                        // Cannot be strictly greater otherwise
-                        // highestNormalizedLiquidityPoolId = 0 if hopTokens[i] balance is 0 in this pool.
-                        highestNormalizedLiquidity
-                    )
-                ) {
-                    highestNormalizedLiquidity = normalizedLiquidity;
-                    highestNormalizedLiquidityPoolId = k;
-                }
-            }
-        }
-        mostLiquidPoolsSecondHop[i] =
-            poolsTokenOutNoTokenIn[highestNormalizedLiquidityPoolId];
-        // console.log(highestNormalizedLiquidity)
-        // console.log(mostLiquidPoolsSecondHop)
-    }
-    return [mostLiquidPoolsFirstHop, mostLiquidPoolsSecondHop, hopTokens];
-}
-
 export const formatSubgraphPools = pools => {
     for (let pool of pools.pools) {
         pool.swapFee = scale(bnum(pool.swapFee), 18);
@@ -916,187 +647,4 @@ export function sortPoolsMostLiquid(
     }
 
     return [mostLiquidPoolsFirstHop, mostLiquidPoolsSecondHop];
-}
-
-export function checkSwapsExactIn(
-    swaps: Swap[][],
-    tokenIn: string,
-    tokenOut: string,
-    amountIn: BigNumber,
-    totalAmtOut: BigNumber,
-    allPoolsNonZeroBalances
-): [Swap[][], BigNumber] {
-    let totalOut = bnum(0);
-
-    for (let i = 0; i < swaps.length; i++) {
-        if (swaps[i].length === 1) {
-            let amtOutFirstSequence = calcOutGivenInForPool(
-                allPoolsNonZeroBalances,
-                swaps[i][0].pool,
-                swaps[i][0].tokenIn,
-                swaps[i][0].tokenOut,
-                bnum(swaps[i][0].swapAmount)
-            );
-            totalOut = totalOut.plus(amtOutFirstSequence);
-        } else {
-            let amtOutFirstSequence = calcOutGivenInForPool(
-                allPoolsNonZeroBalances,
-                swaps[i][0].pool,
-                swaps[i][0].tokenIn,
-                swaps[i][0].tokenOut,
-                bnum(swaps[i][0].swapAmount)
-            );
-
-            if (!amtOutFirstSequence.eq(bnum(swaps[i][1].swapAmount))) {
-                swaps[i][1].swapAmount = amtOutFirstSequence.toString();
-            }
-
-            let amtOutSecondSequence = calcOutGivenInForPool(
-                allPoolsNonZeroBalances,
-                swaps[i][1].pool,
-                swaps[i][1].tokenIn,
-                swaps[i][1].tokenOut,
-                bnum(swaps[i][1].swapAmount)
-            );
-            totalOut = totalOut.plus(amtOutSecondSequence);
-        }
-    }
-
-    if (!totalOut.eq(totalAmtOut)) {
-        return [swaps, totalOut];
-    }
-
-    return [swaps, totalAmtOut];
-}
-
-function calcOutGivenInForPool(
-    Pools,
-    PoolAddr: string,
-    TokenIn: string,
-    TokenOut: string,
-    AmtIn
-) {
-    const swapPool = Pools.pools.find(p => p.id === PoolAddr);
-
-    let pool: PoolPairData = parsePoolPairData(swapPool, TokenIn, TokenOut);
-
-    const amtOut = calcOutGivenIn(
-        pool.balanceIn,
-        pool.weightIn,
-        pool.balanceOut,
-        pool.weightOut,
-        AmtIn,
-        pool.swapFee
-    );
-
-    return amtOut;
-}
-
-export function checkSwapsExactOut(
-    swaps: Swap[][],
-    tokenIn: string,
-    tokenOut: string,
-    amountOut: BigNumber,
-    totalAmtIn: BigNumber,
-    allPoolsNonZeroBalances
-): [Swap[][], BigNumber] {
-    let totalOut = bnum(0);
-    let totalIn = bnum(0);
-
-    for (let i = 0; i < swaps.length; i++) {
-        if (swaps[i].length === 1) {
-            totalOut = totalOut.plus(swaps[i][0].swapAmount);
-            let amtInFirstSequence = calcInGivenOutForPool(
-                allPoolsNonZeroBalances,
-                swaps[i][0].pool,
-                swaps[i][0].tokenIn,
-                swaps[i][0].tokenOut,
-                bnum(swaps[i][0].swapAmount)
-            );
-            totalIn = totalIn.plus(amtInFirstSequence);
-        } else {
-            let amtInSecondSequence = calcInGivenOutForPool(
-                allPoolsNonZeroBalances,
-                swaps[i][1].pool,
-                swaps[i][1].tokenIn,
-                swaps[i][1].tokenOut,
-                swaps[i][1].swapAmount
-            );
-
-            // Amount out of first swap which is input to second swap
-            if (!amtInSecondSequence.eq(bnum(swaps[i][0].swapAmount))) {
-                swaps[i][0].swapAmount = amtInSecondSequence.toString();
-            }
-
-            totalOut = totalOut.plus(swaps[i][1].swapAmount);
-
-            let amtInFirstSequence = calcInGivenOutForPool(
-                allPoolsNonZeroBalances,
-                swaps[i][0].pool,
-                swaps[i][0].tokenIn,
-                swaps[i][0].tokenOut,
-                bnum(swaps[i][0].swapAmount)
-            ); // Swap amount is amount out
-
-            totalIn = totalIn.plus(amtInFirstSequence);
-        }
-    }
-
-    if (!totalOut.eq(amountOut)) {
-        console.log(
-            `Error With Amount Out During Correction - Defaulting To No Swaps`
-        );
-        swaps = [];
-        return [swaps, bnum(0)];
-    }
-
-    if (!totalIn.eq(totalAmtIn)) {
-        return [swaps, totalIn];
-    }
-
-    return [swaps, totalAmtIn];
-}
-
-export function calcInGivenOutForPool(
-    Pools,
-    PoolAddr: string,
-    TokenIn: string,
-    TokenOut: string,
-    AmtIn
-) {
-    const swapPool = Pools.pools.find(p => p.id === PoolAddr);
-
-    let pool: PoolPairData = parsePoolPairData(swapPool, TokenIn, TokenOut);
-
-    const amtOut = calcInGivenOut(
-        pool.balanceIn,
-        pool.weightIn,
-        pool.balanceOut,
-        pool.weightOut,
-        AmtIn,
-        pool.swapFee
-    );
-
-    return amtOut;
-}
-
-export function getPoolsFromSwaps(
-    Swaps: Swap[][],
-    subGraphPools: SubGraphPools
-): SubGraphPools {
-    let pools: SubGraphPools = { pools: [] };
-
-    Swaps.forEach(swapSeq => {
-        if (swapSeq.length === 1) {
-            let pool = subGraphPools.pools.find(p => p.id === swapSeq[0].pool);
-            pools.pools.push(pool);
-        } else {
-            let pool = subGraphPools.pools.find(p => p.id === swapSeq[0].pool);
-            pools.pools.push(pool);
-            pool = subGraphPools.pools.find(p => p.id === swapSeq[1].pool);
-            pools.pools.push(pool);
-        }
-    });
-
-    return pools;
 }
