+++ src/sor.ts
@@ -601,107 +601,3 @@ function getExactSwapAmounts(
     });
     return swapAmounts;
 }
-
-// //// TODO Remove: to debug only!
-// function printSpotPricePathBeforeAndAfterSwap(
-//     path: Path,
-//     swapType: string,
-//     swapAmount: BigNumber
-// ) {
-//     console.log(path.id);
-//     console.log('spotPrice BEFORE trade');
-//     console.log(getSpotPricePath(path).toString());
-
-//     let pathAfterTrade: Path;
-//     pathAfterTrade = path;
-//     if (path.poolPairDataList.length == 1) {
-//         if (swapType === 'swapExactIn') {
-//             path.poolPairDataList[0].balanceIn = path.poolPairDataList[0].balanceIn.plus(
-//                 swapAmount
-//             );
-//             path.poolPairDataList[0].balanceOut = path.poolPairDataList[0].balanceOut.minus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[0],
-//                     swapType,
-//                     swapAmount
-//                 )
-//             );
-//         } else {
-//             path.poolPairDataList[0].balanceIn = path.poolPairDataList[0].balanceIn.plus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[0],
-//                     swapType,
-//                     swapAmount
-//                 )
-//             );
-//             path.poolPairDataList[0].balanceOut = path.poolPairDataList[0].balanceOut.minus(
-//                 swapAmount
-//             );
-//         }
-//     } else {
-//         if (swapType === 'swapExactIn') {
-//             path.poolPairDataList[0].balanceIn = path.poolPairDataList[0].balanceIn.plus(
-//                 swapAmount
-//             );
-//             path.poolPairDataList[0].balanceOut = path.poolPairDataList[0].balanceOut.minus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[0],
-//                     swapType,
-//                     swapAmount
-//                 )
-//             );
-
-//             path.poolPairDataList[1].balanceIn = path.poolPairDataList[1].balanceIn.plus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[0],
-//                     swapType,
-//                     swapAmount
-//                 )
-//             );
-//             path.poolPairDataList[1].balanceOut = path.poolPairDataList[1].balanceOut.minus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[1],
-//                     swapType,
-//                     getReturnAmountSwap(
-//                         path.poolPairDataList[0],
-//                         swapType,
-//                         swapAmount
-//                     )
-//                 )
-//             );
-//         } else {
-//             path.poolPairDataList[0].balanceIn = path.poolPairDataList[0].balanceIn.plus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[0],
-//                     swapType,
-//                     getReturnAmountSwap(
-//                         path.poolPairDataList[1],
-//                         swapType,
-//                         swapAmount
-//                     )
-//                 )
-//             );
-//             path.poolPairDataList[0].balanceOut = path.poolPairDataList[0].balanceOut.minus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[1],
-//                     swapType,
-//                     swapAmount
-//                 )
-//             );
-
-//             path.poolPairDataList[1].balanceIn = path.poolPairDataList[1].balanceIn.plus(
-//                 getReturnAmountSwap(
-//                     path.poolPairDataList[1],
-//                     swapType,
-//                     swapAmount
-//                 )
-//             );
-//             path.poolPairDataList[1].balanceOut = path.poolPairDataList[1].balanceOut.minus(
-//                 swapAmount
-//             );
-//         }
-//     }
-
-//     console.log('spotPrice AFTER  trade');
-//     console.log(getSpotPricePath(path).toString());
-// }
