+++ src/subgraph.ts
@@ -8,87 +8,6 @@ const SUBGRAPH_URL =
     process.env.REACT_APP_SUBGRAPH_URL ||
     'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer';
 
-// LEGACY FUNCTION - Keep Input/Output Format
-export async function getPoolsWithTokens(tokenIn, tokenOut) {
-    // GraphQL is case-sensitive
-    // Always use checksum addresses
-    tokenIn = utils.getAddress(tokenIn);
-    tokenOut = utils.getAddress(tokenOut);
-
-    const query = `
-      query ($tokens: [Bytes!]) {
-          pools (first: 1000, where: {tokensList_contains: $tokens, publicSwap: true, active: true}) {
-            id
-            publicSwap
-            swapFee
-            totalWeight
-            tokensList
-            tokens {
-              id
-              address
-              balance
-              decimals
-              symbol
-              denormWeight
-            }
-          }
-        }
-    `;
-
-    const variables = {
-        tokens: [tokenIn, tokenOut],
-    };
-
-    const response = await fetch(SUBGRAPH_URL, {
-        method: 'POST',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-            query,
-            variables,
-        }),
-    });
-
-    const { data } = await response.json();
-    return data;
-}
-
-// LEGACY FUNCTION - Keep Input/Output Format
-export async function getTokenPairs(token) {
-    // GraphQL is case-sensitive
-    // Always use checksum addresses
-    token = utils.getAddress(token);
-
-    const query = `
-      query ($token: [Bytes!]) {
-          pools (first: 1000, where: {tokensList_contains: $token, publicSwap: true, active: true}) {
-            tokensList
-          }
-        }
-    `;
-
-    const variables = {
-        token: [token],
-    };
-
-    const response = await fetch(SUBGRAPH_URL, {
-        method: 'POST',
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-            query,
-            variables,
-        }),
-    });
-
-    const { data } = await response.json();
-    return data;
-}
-
 // Returns all public pools
 export async function getAllPublicSwapPools(SubgraphUrl: string = '') {
     const query = `
