--- test/multi-pools.json
@@ -1,174 +1,0 @@
-{
-    "balancers": [
-        {
-            "id": "0",
-            "balanceIn": "113.27647924758816",
-            "weightIn": "0.45304813724339865",
-            "balanceOut": "13.728223100754338",
-            "weightOut": "0.5469518627566013",
-            "fee": "0.02",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "1",
-            "balanceIn": "64.74808104925471",
-            "weightIn": "0.787690449123813",
-            "balanceOut": "1.7387670104920505",
-            "weightOut": "0.21230955087618697",
-            "fee": "0.005",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "2",
-            "balanceIn": "179.9817000292248",
-            "weightIn": "0.741518808732097",
-            "balanceOut": "6.297353601218075",
-            "weightOut": "0.25848119126790303",
-            "fee": "0.005",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "3",
-            "balanceIn": "131.62678837890564",
-            "weightIn": "0.25940434505957066",
-            "balanceOut": "37.83245198281663",
-            "weightOut": "0.7405956549404293",
-            "fee": "0.015",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "4",
-            "balanceIn": "156.5190574692619",
-            "weightIn": "0.12539073942448198",
-            "balanceOut": "108.13138456160534",
-            "weightOut": "0.874609260575518",
-            "fee": "0.015",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "5",
-            "balanceIn": "165.22582483699236",
-            "weightIn": "0.6492606574726058",
-            "balanceOut": "8.970600250895505",
-            "weightOut": "0.35073934252739425",
-            "fee": "0.01",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "6",
-            "balanceIn": "150.19836617957264",
-            "weightIn": "0.34614190746454876",
-            "balanceOut": "28.65244033936589",
-            "weightOut": "0.6538580925354512",
-            "fee": "0.02",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "7",
-            "balanceIn": "124.71028367311236",
-            "weightIn": "0.19888867816316935",
-            "balanceOut": "50.48309606368082",
-            "weightOut": "0.8011113218368306",
-            "fee": "0.025",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "8",
-            "balanceIn": "116.51720412309281",
-            "weightIn": "0.18431466919692208",
-            "balanceOut": "51.33929983836195",
-            "weightOut": "0.8156853308030779",
-            "fee": "0.015",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "9",
-            "balanceIn": "93.78932255946289",
-            "weightIn": "0.8319646309041777",
-            "balanceOut": "1.905325520820974",
-            "weightOut": "0.16803536909582228",
-            "fee": "0.005",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "10",
-            "balanceIn": "112.92887769547283",
-            "weightIn": "0.8687561734136942",
-            "balanceOut": "1.6925958036115853",
-            "weightOut": "0.13124382658630585",
-            "fee": "0.005",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "11",
-            "balanceIn": "70.36487993651848",
-            "weightIn": "0.7641356643483449",
-            "balanceOut": "2.169678588285753",
-            "weightOut": "0.2358643356516551",
-            "fee": "0.01",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "12",
-            "balanceIn": "147.2876400429572",
-            "weightIn": "0.19748599725646943",
-            "balanceOut": "60.34568358177149",
-            "weightOut": "0.8025140027435306",
-            "fee": "0.01",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "13",
-            "balanceIn": "39.50121017744873",
-            "weightIn": "0.82132253293356",
-            "balanceOut": "0.8558102730684743",
-            "weightOut": "0.17867746706644005",
-            "fee": "0.030000000000000002",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "14",
-            "balanceIn": "175.6179850883318",
-            "weightIn": "0.5631138999959583",
-            "balanceOut": "13.567552501964828",
-            "weightOut": "0.4368861000040417",
-            "fee": "0.02",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "15",
-            "balanceIn": "72.72519837145565",
-            "weightIn": "0.13368837587959848",
-            "balanceOut": "46.78037607763132",
-            "weightOut": "0.8663116241204015",
-            "fee": "0.005",
-            "spotPrice": "0",
-            "slippage": "0"
-        },
-        {
-            "id": "16",
-            "balanceIn": "113.23123162125968",
-            "weightIn": "0.47346777202158635",
-            "balanceOut": "12.472283640687097",
-            "weightOut": "0.5265322279784137",
-            "fee": "0.005",
-            "spotPrice": "0",
-            "slippage": "0"
-        }
-    ]
-}
